[PyTorch] Declare the instantiation of PackedConvWeightsQnnp<2>::prepack
`PackedConvWeightsQnnp<2>::prepack` is referenced by both `quantized::conv_prepack` and fbgemm.cpp. Since `quantized::conv_prepack` is in the same compilation unit as the class template it was fine. However, if we make operator registration selective, the reference from `quantized::conv_prepack` is gone. The reference from fbgemm.cpp is in another compilation unit and there is link error.

To avoid the link error, instantiate the symbol in the cpp file. It should also work to move all implementations to .h file, but to keep the existing code structure and to avoid (small chance of) code bloat, the implementations are kept as is.

Differential Revision: [D24941989](https://our.internmc.facebook.com/intern/diff/D24941989/)

ghstack-source-id: 117123564
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48256